feat: ini indexer it
